+++ src/Sentry.Unity.iOS/IosNativeScopeObserver.cs
@@ -1,38 +1,18 @@
 using System;
 using System.Runtime.InteropServices;
 using Sentry.Extensibility;
+using Sentry.Unity.Json;
 
 namespace Sentry.Unity.iOS
 {
     public class IosNativeScopeObserver : IScopeObserver
     {
-        [DllImport("__Internal")]
-        private static extern bool CrashedLastRun();
-
-        [DllImport("__Internal")]
-        private static extern void SentryNativeBridgeAddBreadcrumb(string timestamp, string? message, string? type, string? category, int level);
-
-        [DllImport("__Internal")]
-        private static extern void SentryNativeBridgeSetExtra(string key, string? value);
-
-        [DllImport("__Internal")]
-        private static extern void SentryNativeBridgeSetTag(string key, string value);
-
-        [DllImport("__Internal")]
-        private static extern void SentryNativeBridgeUnsetTag(string key);
-
-        [DllImport("__Internal")]
-        private static extern void SentryNativeBridgeSetUser(string? email, string? userId, string? ipAddress, string? username);
-
-        [DllImport("__Internal")]
-        private static extern void SentryNativeBridgeUnsetUser();
-
         private readonly SentryUnityOptions _options;
 
         public IosNativeScopeObserver(SentryUnityOptions options)
         {
             _options = options;
-            _options.CrashedLastRun = CrashedLastRun;
+            _options.CrashedLastRun = SentryCocoaBridgeProxy.CrashedLastRun;
         }
 
         public void AddBreadcrumb(Breadcrumb breadcrumb)
@@ -41,7 +21,7 @@ namespace Sentry.Unity.iOS
             var level = GetBreadcrumbLevel(breadcrumb.Level);
 
             _options.DiagnosticLogger?.LogDebug("Native bridge - Adding breadcrumb m:\"{0}\" l:\"{1}\"", breadcrumb.Message, level);
-            SentryNativeBridgeAddBreadcrumb(timestamp, breadcrumb.Message, breadcrumb.Type, breadcrumb.Category, level);
+            SentryCocoaBridgeProxy.SentryNativeBridgeAddBreadcrumb(timestamp, breadcrumb.Message, breadcrumb.Type, breadcrumb.Category, level);
         }
 
         internal static string GetTimestamp(DateTimeOffset timestamp) =>
@@ -66,7 +46,7 @@ namespace Sentry.Unity.iOS
             string? extraValue = null;
             if (value is not null)
             {
-                extraValue = SerializeExtraValue(value);
+                extraValue = SafeSerializer.SerializeSafely(value);
                 if (extraValue is null)
                 {
                     return;
@@ -74,32 +54,19 @@ namespace Sentry.Unity.iOS
             }
 
             _options.DiagnosticLogger?.LogDebug("Native bridge - Setting Extra k:\"{0}\" v:\"{1}\"", key, value);
-            SentryNativeBridgeSetExtra(key, extraValue);
-        }
-
-        internal string? SerializeExtraValue(object value)
-        {
-            try
-            {
-                return System.Text.Json.JsonSerializer.Serialize(value);
-            }
-            catch (Exception e)
-            {
-                _options.DiagnosticLogger?.LogError("Native bridge - Failed to serialize extra value of type \"{0}\"", e, value.GetType());
-                return null;
-            }
+            SentryCocoaBridgeProxy.SentryNativeBridgeSetExtra(key, extraValue);
         }
 
         public void SetTag(string key, string value)
         {
             _options.DiagnosticLogger?.LogDebug("Native bridge - Setting Tag k:\"{0}\" v:\"{1}\"", key, value);
-            SentryNativeBridgeSetTag(key, value);
+            SentryCocoaBridgeProxy.SentryNativeBridgeSetTag(key, value);
         }
 
         public void UnsetTag(string key)
         {
             _options.DiagnosticLogger?.LogDebug("Native bridge - Unsetting Tag k:\"{0}\"", key);
-            SentryNativeBridgeUnsetTag(key);
+            SentryCocoaBridgeProxy.SentryNativeBridgeUnsetTag(key);
         }
 
         public void SetUser(User? user)
@@ -107,12 +74,12 @@ namespace Sentry.Unity.iOS
             if (user is null)
             {
                 _options.DiagnosticLogger?.LogDebug("Native bridge - Unsetting User");
-                SentryNativeBridgeUnsetUser();
+                SentryCocoaBridgeProxy.SentryNativeBridgeUnsetUser();
             }
             else
             {
                 _options.DiagnosticLogger?.LogDebug("Native bridge - Setting User i:\"{0}\" n:\"{1}\"", user.Id, user.Username);
-                SentryNativeBridgeSetUser(user.Email, user.Id, user.IpAddress, user.Username);
+                SentryCocoaBridgeProxy.SentryNativeBridgeSetUser(user.Email, user.Id, user.IpAddress, user.Username);
             }
         }
     }